Added the getAllCharacter, getCharacterById and getCharactersForSale. But its untested. Also the user tokens in thunderclient are giving errors.

+++ back-end/src/main/java/com/rpg_game/game/controller/AuthController.java
@@ -91,7 +91,7 @@ public class AuthController {
             
             String token = jwtEncoder.encode(JwtEncoderParameters.from(header, claims)).getTokenValue();
 
-            Integer playerId = playerService.findByUsername(authentication.getName())
+            Long playerId = playerService.findByUsername(authentication.getName())
                                                 .map(Player::getId)
                                                 .orElse(null);
